+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 #### 💥 Breaking
 
+- Removed `--include-global` and `--only-local` flags from all applicable commands. Use the new `--config-mode` instead.
 - WASM API
   - Removed the `is_musl` function. Use the host environment instead.